\docType{data}
\name{darwincore_namestypes}
\alias{darwincore_namestypes}
\title{Darwin core terms names and types}
\format{A data frame with 156 rows and 3 variables}
\description{
  A dataset containing the listing of names, short names
  and data types of all 150+ terms are as follows:
}
\details{
  \itemize{ \item name. name of the term \item short_name.
  short name of the term \item type. data type of the term
  like sttring, int etc. }
}
\keyword{datasets}

