\docType{data}
\name{darwincorenames}
\alias{darwincorenames}
\title{Darwin core terms names}
\format{A data frame with 156 rows and 2 variables}
\description{
  A dataset containing the listing of names and short names
  of all 150+ terms are as follows:
}
\details{
  \itemize{ \item name. name of the term \item short_name.
  short name of the term }
}
\keyword{datasets}

