% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{dcdatatypes}
\alias{dcdatatypes}
\title{Darwin core data types}
\format{A data frame with 156 rows and 2 variables}
\usage{
data(dcdatatypes)
}
\description{
A dataset containing the listing of names and data types
of all 150+ terms are as follows:
}
\details{
\itemize{
  \item name. name of the term
  \item type. data type of the term like sttring, int etc.
}
}
\keyword{datasets}

