% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{dcnames}
\alias{dcnames}
\title{Darwin core term names}
\format{A data frame with 156 rows and 2 variables}
\usage{
data(dcnames)
}
\description{
A dataset containing the listing of names and short names
of all 150+ terms are as follows:
}
\details{
\itemize{
  \item name. name of the term
  \item short_name. short name of the term
}
}
\keyword{datasets}

