% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertdump.R
\name{dump-defunct}
\alias{dump-defunct}
\alias{dump_init}
\alias{dump_tbl}
\alias{dump_links}
\title{These functions are defunct}
\usage{
dump_init(...)

dump_tbl(...)

dump_links(...)
}
\description{
These functions are defunct
}
\details{
The Vertnet dumps are too complicated to setup. Please get in
touch with Vertnet or maintainers of this package if you want help using
bulk data.
}
\keyword{internal}
