% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{rvertnet-defunct}
\alias{rvertnet-defunct}
\title{Defunct functions in rvertnet}
\description{
\itemize{
 \item \code{\link{vertavailablemaps}}: Function is now defunct, i.e., not available anymore.
 \item \code{\link{vertlocations}}: Function is now defunct, i.e., not available anymore.
 \item \code{\link{vertoccurrence}}: Function is now defunct, i.e., not available anymore.
 \item \code{\link{vertoccurrencecount}}: Function is now defunct, i.e., not available anymore.
 \item \code{\link{vertproviders}}: Function is now defunct, i.e., not available anymore.
 \item \code{\link{verttaxa}}: Function is now defunct, i.e., not available anymore.
}
}

