% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_dwc_terms}
\alias{simple_dwc_terms}
\title{Darwin core terms}
\format{
\subsection{\code{simple_dwc_terms}}{

A character vector
}
}
\source{
\url{https://raw.githubusercontent.com/tdwg/dwc/master/dist/simple_dwc_vertical.csv}
}
\usage{
simple_dwc_terms
}
\description{
Used internally by \code{\link[=vert_id]{vert_id()}} to filter data to Darwin core terms only. Get in touch with us if these terms need correcting or are out of date.
}
\keyword{datasets}
