% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spatialsearch}
\alias{spatialsearch}
\title{Find records within some distance of a point given latitude and longitude.}
\usage{
spatialsearch(lat = NULL, long = NULL, radius = NULL, lim = 1000,
  compact = TRUE)
}
\arguments{
\item{lat}{Latitude of the central point, in decimal degrees (numeric)}

\item{long}{Longitude of the central point, in decimal degrees (numeric)}

\item{radius}{Radius to search, in meters (numeric)}

\item{lim}{Limit on the number of records returned (numeric)}

\item{compact}{Return a compact data frame (boolean)}
}
\value{
A data frame of search results
}
\description{
Searches by decimal latitude and longitude to return any occurrence record
within the input distance (radius) of the input point.
}
\details{
\code{spatialsearch} finds all records of any taxa having decimal lat/long
   coordinates within a given radius (in meters) of your coordinates.
}
\examples{
\dontrun{

res <- spatialsearch(lat = 33.529, lon = -105.694, rad = 2000, lim = 10)
}
}

