% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatialsearch.R
\name{spatialsearch}
\alias{spatialsearch}
\title{Find records within some distance of a point given latitude and longitude.}
\usage{
spatialsearch(lat, long, radius, limit = 1000, compact = TRUE,
  verbose = TRUE)
}
\arguments{
\item{lat}{Latitude of the central point, in decimal degrees (numeric) Required.}

\item{long}{Longitude of the central point, in decimal degrees (numeric) Required.}

\item{radius}{Radius to search, in meters (numeric). Required.}

\item{limit}{Limit on the number of records returned (numeric). Default: 1000}

\item{compact}{Return a compact data frame (logical)}

\item{verbose}{Print progress and information messages. Default: TRUE}
}
\value{
A data frame of search results
}
\description{
Searches by decimal latitude and longitude to return any occurrence record
within the input distance (radius) of the input point.
}
\details{
\code{\link{spatialsearch}} finds all records of any taxa having decimal lat/long
   coordinates within a given radius (in meters) of your coordinates.
}
\examples{
\dontrun{
res <- spatialsearch(lat = 33.529, long = -105.694, radius = 2000, limit = 10)
}
}
\references{
\url{https://github.com/VertNet/webapp/wiki/The-API-search-function}
}

