% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatialsearch.R
\name{spatialsearch}
\alias{spatialsearch}
\title{Find records within some distance of a point given latitude and longitude.}
\usage{
spatialsearch(lat, long, radius, limit = 1000, compact = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{lat}{Latitude of the central point, in decimal degrees (numeric) Required.}

\item{long}{Longitude of the central point, in decimal degrees (numeric) Required.}

\item{radius}{Radius to search, in meters (numeric). There is no default value for
this parameter. Required.}

\item{limit}{Limit on the number of records returned. If you use a number with 5 zeros or more
(e.g., 100000), you need to pass it as \code{100000L}. Default: 1000. (numeric)}

\item{compact}{Return a compact data frame (logical)}

\item{verbose}{Print progress and information messages. Default: TRUE}

\item{...}{Curl arguments passed on to \code{\link[httr]{GET}}}
}
\value{
A data frame of search results
}
\description{
Searches by decimal latitude and longitude to return any occurrence record
within the input distance (radius) of the input point.
}
\details{
\code{\link{spatialsearch}} finds all records of any taxa having decimal lat/long
   coordinates within a given radius (in meters) of your coordinates.
}
\examples{
\dontrun{
res <- spatialsearch(lat = 33.529, long = -105.694, radius = 2000, limit = 10)

# Pass in curl options for curl debugging
library("httr")
out <- spatialsearch(lat = 33.529, long = -105.694, radius = 2000, limit = 10, config=verbose())
}
}
\references{
\url{https://github.com/VertNet/webapp/wiki/The-API-search-function}
}

