% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vertsearch.R
\name{vertsearch}
\alias{vertsearch}
\title{Find records using a global full-text search of VertNet archives.}
\usage{
vertsearch(taxon = NULL, ..., limit = 1000, compact = TRUE,
  verbose = TRUE)
}
\arguments{
\item{taxon}{Taxonomic identifier or other text to search for (character)}

\item{...}{Additional search terms (character)}

\item{limit}{Limit on the number of records returned; up to 1000 (numeric)}

\item{compact}{Return a compact data frame (boolean)}

\item{verbose}{Print progress and information messages. Default: TRUE}
}
\value{
A data frame of search results
}
\description{
Returns any record containing your target text in any field of the record.
}
\details{
\code{\link{vertsearch}} performs a nonspecific search for your input within
   every record and field of the VertNet archives. For a more specific
   search, try searchbyterm().
}
\examples{
\dontrun{
out <- vertsearch(taxon = "aves", state = "california")

# Limit the number of records returned (under 1000)
out <- vertsearch("(kansas state OR KSU)", lim = 200)
# Use bigsearch() to retrieve >1000 records

# Find multiple species using searchbyterm():
# a) returns a specific result
out <- searchbyterm(gen = "mustela", sp = "(nivalis OR erminea)")
vertmap(out)

# b) returns a non-specific result
out <- vertsearch(tax = "(mustela nivalis OR mustela erminea)")
vertmap(out)

# c) returns a non-specific result
splist <- c("mustela nivalis", "mustela erminea")
out <- lapply(splist, function(x) vertsearch(t=x, lim=500))
vertmap(out)
}
}
\references{
\url{https://github.com/VertNet/webapp/wiki/The-API-search-function}
}

